Initial attempt at a global time range slider. Contains unwanted side effects.
The slider can update graphs currently being displayed, but newly
added graphs are not currently updated. This means that the slider
needs to be used again after a new graph is added. Additionally, since
we are now storing the displayed graphs, switching scenarios does not currently
clear them. Eventually this may be desirable, as we may want to enable
being able to compare two of the "same" graphs across different scenarios,
but the way the dropdown currently works prevents this.